add routing by order number show api validation errors per field change WeightKg from double to decimal add AsNoTracking to read queries handle OperationCanceledException in middleware store pagination page in URL query params

+++ src/DeliveryWebApp.API/Application/DTOs/CreateOrderDto.cs
@@ -23,7 +23,7 @@ public class CreateOrderDto
 
     [Required(ErrorMessage = "Вес груза обязателен")]
     [Range(0.1, 10000, ErrorMessage = "Вес должен быть от 0.1 до 10000 кг")]
-    public double WeightKg { get; init; }
+    public decimal WeightKg { get; init; }
 
     [Required(ErrorMessage = "Дата забора груза обязательна")]
     [FutureDate]
